added postprocessing with ffmpeg for rescaling

+++ download.py
@@ -1,7 +1,6 @@
 import argparse
 import json
 import shutil
-import subprocess
 import uuid
 import warnings
 from collections import defaultdict
@@ -12,72 +11,56 @@ import pandas as pd
 from joblib import delayed
 from tqdm import tqdm
 
-from utils import ProgressParallel
+from postprocess import compress_video, VideoPreset
+from utils import ProgressParallel, apply_subprocess
 
 
 class VideoClipInfo(NamedTuple):
     video_id: str
-    start_time: int
-    end_time: int
+    start_time: Optional[int]
+    end_time: Optional[int]
     split: Optional[str]
     label: Optional[str]
 
 
 def download_video(video_id: str,
                    output_filename: Path,
-                   start_time: int,
-                   end_time: int,
+                   start_time: Optional[int] = None,
+                   end_time: Optional[int] = None,
                    num_retries: int = 5,
                    post_process_args: Optional[Dict] = None,
                    url_base: str = 'https://www.youtube.com/watch?v=',
                    ) -> Tuple[bool, str]:
     assert len(video_id) == 11, f'video_identifier must have length 11 but got {video_id}'
-    status = False
 
     if post_process_args is None:
         tmp_filename = output_filename
     else:
         tmp_filename = (output_filename.parent.parent / 'tmp') / f'{uuid.uuid4()}.mp4'
 
+    section = f'--download-sections *{start_time}-{end_time}' if start_time is not None and end_time is not None else ''
+
     command = ['yt-dlp',
                '--quiet',
                '--no-warnings',
                '--retries', str(num_retries),
                '-f', 'mp4',
-               '--download-sections', f'*{start_time}-{end_time}',
+               section,
                '-o', f'{tmp_filename}',
                f'{url_base + video_id}']
-    try:
-        subprocess.check_output(command, shell=False, stderr=subprocess.STDOUT)
-    except subprocess.CalledProcessError as error:
-        return status, str(error.output)
+    downloaded, log = apply_subprocess(command)
 
-    if post_process_args is not None:
-        default = dict(crf='18', preset='veryslow', width='224', height='224')
+    if post_process_args is not None and downloaded:
+        default = dict(crf=28, preset=VideoPreset.medium, width=224, height=224, fps=30, sample_rate=16000, channels=1)
         default.update(post_process_args)
+        post_success, log = compress_video(tmp_filename, output_filename, **default)
 
-        command = ['ffmpeg',
-                   '-i', f'"{tmp_filename}"',
-                   '-vf', f"scale={default['width']}:{default['height']}",
-                   '-c:v', 'libx264',
-                   '-crf', str(default['crf']),
-                   '-preset', str(default['preset']),
-                   '-c:a', 'copy',
-                   '-threads', '1',
-                   '-loglevel', 'panic',
-                   f'"{output_filename}"']
-        command = ' '.join(command)
-
-        try:
-            subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
-        except subprocess.CalledProcessError as error:
-            return status, str(error.output)
-
+        # remove tmp file
         tmp_filename.unlink()
+        return post_success, log
 
     status = output_filename.exists()
-
-    return status, 'Downloaded'
+    return status, log
 
 
 def download_wrapper(row: VideoClipInfo,
@@ -89,7 +72,7 @@ def download_wrapper(row: VideoClipInfo,
     filename = get_video_filename(row, label_to_dir, by, time_format)
 
     if filename.exists():
-        return row.video_id, True, 'Exists'
+        return row.video_id, True, 'exists'
 
     downloaded, log = download_video(row.video_id, filename, row.start_time, row.end_time, num_retries,
                                      post_process_args)
@@ -105,8 +88,9 @@ def parse_input(csv_path: Union[str, Path]) -> pd.DataFrame:
         ('time_end', 'end_time'),
     ])
     df.rename(columns=columns, inplace=True)
-    df['start_time'] = df['start_time'].apply(int)
-    df['end_time'] = df['end_time'].apply(int)
+    df['start_time'] = df['start_time'].apply(int) if 'start_time' in df.columns else None
+    df['end_time'] = df['end_time'].apply(int) if 'end_time' in df.columns else None
+
     return df
 
 
@@ -117,7 +101,7 @@ def create_dir(dataset: pd.DataFrame,
     data_dir = output_dir / 'data'
     data_dir.mkdir(exist_ok=True)
 
-    (output_dir / 'tmp').mkdir(exist_ok=False)
+    (output_dir / 'tmp').mkdir(exist_ok=True)
 
     label_to_dir = defaultdict(lambda: data_dir)
 
@@ -137,11 +121,13 @@ def get_video_filename(row: VideoClipInfo,
                        label_to_dir: defaultdict[Path],
                        by: Optional[str] = None,
                        time_format: str = '06d') -> Path:
-    filename = f"{row.video_id}_{row.start_time:{time_format}}_{row.end_time:{time_format}}.mp4"
-    if by is None:
-        base = label_to_dir.default_factory()
+    if row.start_time is None and row.end_time is None:
+        filename = f"{row.video_id}.mp4"
     else:
-        base = label_to_dir[getattr(row, by)]
+        filename = f"{row.video_id}_{row.start_time:{time_format}}_{row.end_time:{time_format}}.mp4"
+
+    base = label_to_dir.default_factory() if by is None else label_to_dir[getattr(row, by)]
+
     return base / filename
 
 
@@ -151,7 +137,7 @@ def main(csv_path: str,
          time_format: str = '06d',
          retries: int = 5,
          n_jobs: int = 32,
-         post_process_args: Optional[Dict] = dict(crf='18', preset='veryslow', width='224', height='224'),
+         post_process_args: Optional[Dict] = dict(),  # dict(crf='18', preset='veryslow', width='224', height='224'),
          verbose: bool = True) -> List:
     csv_path = Path(csv_path)
     output_dir = Path(output_dir)
@@ -208,6 +194,5 @@ if __name__ == '__main__':
                         help='Number of retries for downloading one video')
     parser.add_argument('-q', '--quiet', action='store_false', dest='verbose',
                         help='Prevent the display of a progress bar')
-
 
     main(**vars(parser.parse_args()))
